Sistema de guardado + Main Menu + Pause Menu (en progreso)

+++ Assets/Scripts/PlayerMenu/Inventary/Weapons/Weapon.cs
@@ -23,6 +23,6 @@ public class Weapon : ScriptableObject
     public float ProjectileTimeLife;
 
     [Header("Stats")]
-    public float attack;
-    public float defense;
+    public int attack;
+    public int defense;
 }
